refactor: split compiler.go into domain-specific files
Split pkg/config/compiler.go (5878 lines) into 8 domain-specific files:

- compiler_security.go: compileSecurity, compileZones, compilePolicies, etc.
- compiler_interfaces.go: compileInterfaces, parseMSSValue
- compiler_protocols.go: compileProtocols, compileRouterAdvertisement, etc.
- compiler_ipsec.go: compileIKE, compileIPsec, parseDeadPeerDetectionNode
- compiler_routing.go: compileRoutingOptions, compileStaticRoutes, etc.
- compiler_firewall.go: compileFirewall, compileFilterFrom, compileFilterThen
- compiler_system.go: compileSystem, compileChassis, compileSNMP, etc.
- compiler_services.go: compileDHCPLocalServer, compileRPM, etc.

compiler.go retains top-level dispatch, ValidateConfig, compileApplications,
and shared helpers (nodeVal, normalizeProtocol, validatePortSpec).

Mechanical move only — no logic changes. Follows the pattern established
by compiler_nat.go.

Fixes #545
